+++ vitest.config.ts
@@ -1,6 +1,14 @@
+import { fileURLToPath, URL } from "node:url";
 import { defineConfig } from "vitest/config";
+import vue from "@vitejs/plugin-vue";
 
 export default defineConfig({
+  plugins: [vue()],
+  resolve: {
+    alias: {
+      "@": fileURLToPath(new URL("./src", import.meta.url)),
+    },
+  },
   test: {
     environment: "jsdom",
     coverage: {
@@ -10,6 +18,7 @@ export default defineConfig({
         "**/data/**",
         "**/locales/**",
         "**/constants.ts",
+        "**/main.ts",
         "coverage/**",
         "dist/**",
         "packages/*/test{,s}/**",