+++ CHANGELOG.md
@@ -36,6 +36,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Velites grit to 4
 - Bard is showed now
 
+### Added
+
+- Tests for Utils
+
 ## [v0.20.12]
 
 ### Added